Revert the moving of user-based settings to Core/User namespace and merge site-based settings in the Analytics4 namespace.
