+++ assets/js/modules/analytics-4/components/site-goals/selection-panel/SaveErrorNotice.tsx
@@ -26,7 +26,6 @@ import { __, sprintf } from '@wordpress/i18n';
  */
 import { Select, useSelect } from 'googlesitekit-data';
 import SelectionPanelError from '@/js/components/SelectionPanel/SelectionPanelError';
-import { CORE_USER } from '@/js/googlesitekit/datastore/user/constants';
 import useFormValue from '@/js/hooks/useFormValue';
 import {
 	SITE_GOALS_MAX_SELECTED_DRIVERS,
@@ -40,6 +39,7 @@ import {
 	GoalType,
 } from '@/js/modules/analytics-4/components/site-goals/goal-drivers/types';
 import { getSelectedDriverIDs } from '@/js/modules/analytics-4/components/site-goals/utils/selectedDrivers';
+import { MODULES_ANALYTICS_4 } from '@/js/modules/analytics-4/datastore/constants';
 
 interface SaveErrorNoticeProps {
 	hasEcommerceGoalDrivers: boolean;
@@ -60,7 +60,7 @@ export default function SaveErrorNotice( {
 
 	const saveError = useSelect(
 		( select: Select ) =>
-			select( CORE_USER ).getErrorForAction(
+			select( MODULES_ANALYTICS_4 ).getErrorForAction(
 				'saveSiteGoalsSettings',
 				[]
 			),
