+++ assets/js/modules/analytics-4/components/site-goals/widgets/LeadGenerationPerformanceWidget.tsx
@@ -267,7 +267,8 @@ const LeadGenerationPerformanceWidget: FC<
 		[]
 	);
 	const effectiveSelectedDrivers = useSelect(
-		( select: Select ) => select( CORE_USER ).getSiteGoalsGoalDrivers(),
+		( select: Select ) =>
+			select( MODULES_ANALYTICS_4 ).getSiteGoalsGoalDrivers(),
 		[]
 	) as GoalDriverSelectionState | undefined;
 	const resolvedSelections = resolveGoalDriverSelectionState(
