+++ assets/js/modules/analytics-4/components/site-goals/widgets/OnlineStorePerformanceWidget.tsx
@@ -177,7 +177,8 @@ const OnlineStorePerformanceWidget: FC<
 	);
 
 	const effectiveSelectedDrivers = useSelect(
-		( select: Select ) => select( CORE_USER ).getSiteGoalsGoalDrivers(),
+		( select: Select ) =>
+			select( MODULES_ANALYTICS_4 ).getSiteGoalsGoalDrivers(),
 		[]
 	) as GoalDriverSelectionState | undefined;
 	const resolvedSelections = resolveGoalDriverSelectionState(
@@ -186,7 +187,7 @@ const OnlineStorePerformanceWidget: FC<
 
 	const effectiveVisitorEngagement = useSelect(
 		( select: Select ) =>
-			select( CORE_USER ).getSiteGoalsVisitorEngagement(),
+			select( MODULES_ANALYTICS_4 ).getSiteGoalsVisitorEngagement(),
 		[]
 	);
 	const resolvedVisitorEngagement = resolveVisitorEngagementSelectionState(
