+++ includes/Core/User/User.php
@@ -47,14 +47,6 @@ class User {
 	private $email_reporting;
 
 	/**
-	 * Site_Goals instance.
-	 *
-	 * @since n.e.x.t
-	 * @var Site_Goals
-	 */
-	private $site_goals;
-
-	/**
 	 * Initial_Setup instance.
 	 *
 	 * @since 1.164.0
@@ -68,17 +60,12 @@ class User {
 	 * @since 1.134.0
 	 * @since 1.162.0 Added Email Reporting.
 	 * @since 1.164.0 Added Initial Setup.
-	 * @since n.e.x.t Added Site Goals.
 	 *
 	 * @param User_Options $user_options User_Options instance.
 	 */
 	public function __construct( User_Options $user_options ) {
 		$this->audience_segmentation = new Audience_Segmentation( $user_options );
 		$this->conversion_reporting  = new Conversion_Reporting( $user_options );
-
-		if ( Feature_Flags::enabled( 'siteGoals' ) ) {
-			$this->site_goals = new Site_Goals( $user_options );
-		}
 
 		if ( Feature_Flags::enabled( 'proactiveUserEngagement' ) ) {
 			$this->email_reporting = new Email_Reporting( $user_options );
@@ -95,15 +82,10 @@ class User {
 	 * @since 1.134.0
 	 * @since 1.162.0 Added Email Reporting.
 	 * @since 1.164.0 Added Initial Setup.
-	 * @since n.e.x.t Added Site Goals.
 	 */
 	public function register() {
 		$this->audience_segmentation->register();
 		$this->conversion_reporting->register();
-
-		if ( Feature_Flags::enabled( 'siteGoals' ) && $this->site_goals ) {
-			$this->site_goals->register();
-		}
 
 		if ( Feature_Flags::enabled( 'proactiveUserEngagement' ) && $this->email_reporting ) {
 			$this->email_reporting->register();
