+++ includes/Modules/Analytics_4.php
@@ -94,9 +94,10 @@ use Google\Site_Kit\Modules\Analytics_4\Synchronize_Property;
 use Google\Site_Kit\Modules\Analytics_4\Synchronize_AdSenseLinked;
 use Google\Site_Kit\Modules\Analytics_4\GoogleAnalyticsAdmin\AccountProvisioningService;
 use Google\Site_Kit\Modules\Analytics_4\Resource_Data_Availability_Date;
+use Google\Site_Kit\Modules\Analytics_4\Datapoints\Save_Site_Goals_Settings;
 use Google\Site_Kit\Modules\Analytics_4\Settings;
 use Google\Site_Kit\Modules\Analytics_4\Site_Goals_Settings;
-use Google\Site_Kit\Core\User\Site_Goals_Settings as User_Site_Goals_Settings;
+use Google\Site_Kit\Modules\Analytics_4\Site_Goals_Site_Settings;
 use Google\Site_Kit\Modules\Analytics_4\Synchronize_AdsLinked;
 use Google\Site_Kit\Modules\Analytics_4\Tag_Guard;
 use Google\Site_Kit\Modules\Analytics_4\Tag_Interface;
@@ -198,23 +199,22 @@ final class Analytics_4 extends Module implements Module_With_Inline_Data, Modul
 	protected $audience_settings;
 
 	/**
-	 * Site-wide Site_Goals_Settings instance.
+	 * Per-user Site_Goals_Settings instance.
 	 *
-	 * @since n.e.x.t
+	 * @since 1.181.0
 	 *
 	 * @var Site_Goals_Settings
 	 */
 	protected $site_goals_settings;
 
 	/**
-	 * Per-user Site_Goals_Settings instance.
+	 * Site-wide Site_Goals_Site_Settings instance.
 	 *
-	 * @since 1.181.0
-	 * @since 1.182.0 Renamed from $site_goals_settings and type moved to Core\User.
+	 * @since n.e.x.t
 	 *
-	 * @var User_Site_Goals_Settings
+	 * @var Site_Goals_Site_Settings
 	 */
-	protected $user_site_goals_settings;
+	protected $site_goals_site_settings;
 
 	/**
 	 * Audience_Utilities instance.
@@ -265,8 +265,8 @@ final class Analytics_4 extends Module implements Module_With_Inline_Data, Modul
 		$this->custom_dimensions_data_available  = new Custom_Dimensions_Data_Available( $this->transients );
 		$this->reset_audiences                   = new Reset_Audiences( $this->user_options );
 		$this->audience_settings                 = new Audience_Settings( $this->options );
-		$this->site_goals_settings               = new Site_Goals_Settings( $this->options );
-		$this->user_site_goals_settings          = new User_Site_Goals_Settings( $this->user_options );
+		$this->site_goals_settings               = new Site_Goals_Settings( $this->user_options );
+		$this->site_goals_site_settings          = new Site_Goals_Site_Settings( $this->options );
 		$this->audience_utilities                = new Audience_Utilities( $this->audience_settings );
 		$this->resource_data_availability_date   = new Resource_Data_Availability_Date( $this->transients, $this->get_settings(), $this->audience_settings );
 		$this->advanced_data_breakdowns_settings = new Advanced_Data_Breakdowns_Settings( $this->options );
@@ -311,7 +311,7 @@ final class Analytics_4 extends Module implements Module_With_Inline_Data, Modul
 
 		$this->audience_settings->register();
 		$this->site_goals_settings->register();
-		$this->user_site_goals_settings->register();
+		$this->site_goals_site_settings->register();
 		$this->advanced_data_breakdowns_settings->register();
 
 		( new Advanced_Tracking( $this->context ) )->register();
@@ -596,7 +596,7 @@ final class Analytics_4 extends Module implements Module_With_Inline_Data, Modul
 		$this->custom_dimensions_data_available->reset_data_available();
 		$this->reset_audiences->reset_audience_data();
 		$this->audience_settings->delete();
-		$this->user_site_goals_settings->delete();
+		$this->site_goals_settings->delete();
 		$this->advanced_data_breakdowns_settings->delete();
 	}
 
@@ -987,6 +987,13 @@ final class Analytics_4 extends Module implements Module_With_Inline_Data, Modul
 				)
 			),
 			'GET:site-goals-settings'                   => new Get_Site_Goals_Settings(
+				array(
+					'site_goals_settings'      => $this->site_goals_settings,
+					'site_goals_site_settings' => $this->site_goals_site_settings,
+					'service'                  => '',
+				)
+			),
+			'POST:save-site-goals-settings'             => new Save_Site_Goals_Settings(
 				array(
 					'site_goals_settings' => $this->site_goals_settings,
 					'service'             => '',
