+++ tests/phpunit/integration/Modules/Analytics_4Test.php
@@ -1726,6 +1726,7 @@ class Analytics_4Test extends TestCase {
 				'save-resource-data-availability-date',
 				'sync-audiences',
 				'site-goals-settings',
+				'save-site-goals-settings',
 			),
 			$this->analytics->get_datapoints(),
 			'Analytics 4 module should expose the expected datapoints'
@@ -1765,6 +1766,7 @@ class Analytics_4Test extends TestCase {
 				'save-resource-data-availability-date',
 				'sync-audiences',
 				'site-goals-settings',
+				'save-site-goals-settings',
 			),
 			$this->analytics->get_datapoints(),
 			'Analytics 4 module should expose the expected datapoints with conversion reporting'
